+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/pages/citizen/registration/SelectUserAddress.js
@@ -1,117 +1,40 @@
 import { FormComposerV2 } from "@egovernments/digit-ui-module-core";
-import React, { useEffect, useMemo, useState } from "react";
+import React, { useMemo, useState } from "react";
 import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
 import isEqual from "lodash/isEqual";
-import { getFileByFileStore } from "../../../Utils";
+import {
+  createRegistrationPatternValidationOnChange,
+  syncAddressFormDataIfChanged,
+  useUserRegistrationSessionRestore,
+} from "./shared/registrationFlowShared";
 
 const SelectUserAddress = ({ config, onSelect, t, params, pathOnRefresh }) => {
   const history = useHistory();
   const [isDisabled, setIsDisabled] = useState(false);
   const [formdata, setformData] = useState(params?.address || {});
 
-  const onFormValueChange = (setValue, formData, formState) => {
-    if (formData?.isBothAddressSame?.code === "YES") {
-      const existing = formData?.currentAddress;
-      const target = { ...formData?.addressDetails, temp: 1 };
-      if (!isEqual(existing, target)) {
-        setValue("currentAddress", target);
+  const onFormValueChange = createRegistrationPatternValidationOnChange(config, setIsDisabled, {
+    onAfterValidate: (setValue, formData) => {
+      if (formData?.isBothAddressSame?.code === "YES") {
+        const existing = formData?.currentAddress;
+        const target = { ...formData?.addressDetails, temp: 1 };
+        if (!isEqual(existing, target)) {
+          setValue("currentAddress", target);
+        }
+      } else if (formData?.isBothAddressSame?.code === "NO" && formData?.currentAddress?.temp) {
+        setValue("currentAddress", undefined);
       }
-    } else if (formData?.isBothAddressSame?.code === "NO" && formData?.currentAddress?.temp) {
-      setValue("currentAddress", undefined);
-    }
+      syncAddressFormDataIfChanged(formdata, formData, setformData);
+    },
+  });
 
-    let isDisabled = false;
-    config.forEach((curr) => {
-      if (isDisabled) return;
-      if (!(curr.body[0].key in formData) || !formData[curr.body[0].key]) {
-        return;
-      }
-      curr.body[0].populators.inputs.forEach((input) => {
-        if (isDisabled) return;
-        if (Array.isArray(input.name)) return;
-        if (
-          formData[curr.body[0].key][input.name] &&
-          formData[curr.body[0].key][input.name].length > 0 &&
-          !["documentUpload", "radioButton"].includes(input.type) &&
-          input.validation &&
-          !formData[curr.body[0].key][input.name].match(Digit.Utils.getPattern(input.validation.patternType) || input.validation.pattern)
-        ) {
-          isDisabled = true;
-        }
-        if (Array.isArray(formData[curr.body[0].key][input.name]) && formData[curr.body[0].key][input.name].length === 0) {
-          isDisabled = true;
-        }
-      });
-    });
-    if (isDisabled) {
-      setIsDisabled(isDisabled);
-    } else {
-      setIsDisabled(false);
-    }
-
-    if (!isEqual(formdata, formData)) {
-      setformData(formData);
-    }
-  };
-
-  useEffect(() => {
-    const handleRedirect = async () => {
-      if (!params?.name) {
-        const storedParams = sessionStorage.getItem("userRegistrationParams");
-        let newParams = storedParams ? JSON.parse(storedParams) : params;
-
-        const fileStoreId = newParams?.uploadedDocument?.filedata?.files?.[0]?.fileStoreId;
-        const filename = newParams?.uploadedDocument?.filename;
-
-        const barCouncilFileStoreId = newParams?.formData?.clientDetails?.barCouncilId?.[1]?.fileStoreId;
-        const barCouncilFilename = newParams?.formData?.clientDetails?.barCouncilId?.[0];
-
-        if (barCouncilFileStoreId && barCouncilFilename) {
-          const barCouncilUri = `${
-            window.location.origin
-          }/filestore/v1/files/id?tenantId=${Digit.ULBService.getCurrentTenantId()}&fileStoreId=${barCouncilFileStoreId}`;
-          const barCouncilFile = await getFileByFileStore(barCouncilUri, barCouncilFilename);
-
-          newParams = {
-            ...newParams,
-            formData: {
-              ...newParams.formData,
-              clientDetails: {
-                ...newParams.formData.clientDetails,
-                barCouncilId: [
-                  [
-                    barCouncilFilename,
-                    {
-                      file: barCouncilFile,
-                      fileStoreId: barCouncilFileStoreId,
-                    },
-                  ],
-                ],
-              },
-            },
-          };
-        }
-
-        if (fileStoreId && filename) {
-          const uri = `${window.location.origin}/filestore/v1/files/id?tenantId=${Digit.ULBService.getCurrentTenantId()}&fileStoreId=${fileStoreId}`;
-          const file = await getFileByFileStore(uri, filename);
-
-          newParams = {
-            ...newParams,
-            uploadedDocument: {
-              ...newParams.uploadedDocument,
-              file,
-            },
-          };
-        }
-
-        sessionStorage.removeItem("userRegistrationParams");
-        history.push(pathOnRefresh, { newParams });
-      }
-    };
-
-    handleRedirect();
-  }, [params?.address, params, history, pathOnRefresh]);
+  useUserRegistrationSessionRestore({
+    params,
+    history,
+    pathOnRefresh,
+    shouldRestore: (p) => !p?.name,
+    effectDeps: [params?.address],
+  });
 
   const modifiedConfig = useMemo(() => {
     const shouldDisable = formdata?.isBothAddressSame?.code === "YES";
