+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/orders/src/pages/employee/shared/paymentSummonPostReceiptDefaults.js
@@ -15,3 +15,16 @@ export const submitModalInfoPostPayment = {
   isArrow: false,
   showTable: true,
 };
+
+export const submitModalInfoRpadPayment = {
+  header: "CS_HEADER_FOR_SUMMON_RPAD",
+  subHeader: "CS_SUBHEADER_TEXT_FOR_Summon_RPAD",
+  caseInfo: [
+    {
+      key: "Case Number",
+      value: "FSM-2019-04-23-898898",
+    },
+  ],
+  isArrow: false,
+  showTable: true,
+};
